+++ src/api/authApi.ts
@@ -1,48 +1,33 @@
-/* eslint-disable no-useless-catch */
+ 
 
-import { ProfileType, SignInResponse, SignInTypes, SignUpTypes } from "../types";
+import {
+	ProfileType,
+	SignInResponse,
+	SignInTypes,
+	SignUpTypes,
+} from "../types/authTypes";
 import { instance, tokensInstance } from "./axiosInstance";
 
 export async function signUpApi(userData: SignUpTypes): Promise<ProfileType> {
-	try {
-		const res = await instance.post("/auth/signup", userData);
-		return res.data;
-	} catch (error) {
-		throw error;
-	}
+	const res = await instance.post("/auth/signup", userData);
+	return res.data;
 }
 
 export async function signInApi(userData: SignInTypes): Promise<SignInResponse> {
-	try {
-		const res = await tokensInstance.post("/auth/signin", userData);
-		return res.data;
-	} catch (error) {
-		throw error;
-	}
+	const res = await tokensInstance.post("/auth/signin", userData);
+	return res.data;
 }
 
 export async function refreshToken(refreshToken: string): Promise<SignInResponse> {
-	try {
-		const res = await tokensInstance.post("/auth/refresh", { refreshToken });
-		return res.data;
-	} catch (error) {
-		throw error;
-	}
+	const res = await tokensInstance.post("/auth/refresh", { refreshToken });
+	return res.data;
 }
 
 export async function getProfile(): Promise<ProfileType> {
-	try {
-		const res = await instance.get("/user/profile");
-		return res.data;
-	} catch (error) {
-		throw error;
-	}
+	const res = await instance.get("/user/profile");
+	return res.data;
 }
 
 export async function logoutApi(): Promise<void> {
-	try {
-		await instance.post("/user/logout");
-	} catch (error) {
-		throw error;
-	}
+	await instance.post("/user/logout");
 }
