--- src/types.ts
@@ -1,93 +1,0 @@
-export interface TodoRequest {
-	title?: string;
-	isDone?: boolean; // изменение статуса задачи происходит через этот флаг
-}
-
-export interface Todo {
-	id: number;
-	title: string;
-	created: string;
-	isDone: boolean;
-}
-
-export interface TodoInfo {
-	all: number;
-	completed: number;
-	inWork: number;
-}
-
-export interface MetaResponse<T, N> {
-	data: T[];
-	info?: N;
-	meta: {
-		totalAmount: number;
-	};
-}
-
-export interface Task {
-	created: string;
-	id: number;
-	isDone: boolean;
-	title: string;
-}
-
-export enum TodoFilterEnum {
-	all = "all",
-	inWork = "inWork",
-	completed = "completed",
-}
-
-export interface IsLoggedIn {
-	isLogged: boolean;
-}
-
-export interface SignInTypes {
-	login: string;
-	password: string;
-}
-
-export interface SignInResponse {
-	accessToken: string;
-	refreshToken: string;
-}
-
-export interface ProfileType {
-	date: string;
-	email: string;
-	id: number;
-	isAdmin: boolean;
-	isBlocked: boolean;
-	phoneNumber: string;
-	username: string;
-}
-
-export interface SignUpTypes {
-	email: string;
-	login: string;
-	password: string;
-	phoneNumber: string;
-	username: string;
-}
-
-export interface SignUpResponse {
-	date: string;
-	email: string;
-	id: number;
-	isAdmin: boolean;
-	isBlocked: boolean;
-	phoneNumber: string;
-	username: string;
-}
-
-export type PersonalOrPublic = "personal" | "public";
-
-export type RequestTypes =
-	| "signUp"
-	| "signIn"
-	| "refreshToken"
-	| "getProfile"
-	| "logout"
-	| "getTasks"
-	| "createTask"
-	| "updateTask"
-	| "deleteTask";
