+++ src/types/commonTypes.ts
@@ -1,0 +1,11 @@
+export type PersonalOrPublic = "personal" | "public";
+
+export type RequestTypes =
+	| "signUp"
+	| "signIn"
+	| "refreshToken"
+	| "logout"
+	| "getTasks"
+	| "createTask"
+	| "updateTask"
+	| "deleteTask";
